projects/weather_data/requirements.txt
async-timeout==4.0.2
black==23.1.0
flake8==6.0.0
Flask==2.2.2
flask-restx==1.0.6
pycodestyle==2.10.0
pyflakes==3.0.1
SQLAlchemy==2.0.3
pyyaml
